Update awsStream.py
Sets up connection to EC2 server that is setup properly. Currently delivers audio and video where the audio is silent if the button is not pressed (to allow the speaker to work on the same I2S rail). Raw HTML can be grabbed from the server that contains the stream.